Reduce uptime proof frequency to 31 minutes
This lowers the SN uptime proof sending interval to 31 minutes, while
keeping the minimum proof time accepted by other SNs at 30 minutes.
This maintains backwards compatibility with the existing SN code, which
already accepts proofs for a SN with an uptime over 30 minutes.

The intent here is to avoid accidental deregistration even for a SN that
is up; there are some edge cases (particularly surrounding a reboot or
SN restarts with particular timing windows) where a SN uptime proof can
fail to go out; this commit makes it a little harder to get
deregistered.