Delay initial uptime proof by 60 seconds
It is quite possible for the initial uptime proof after starting to fail
to go out, in which case the first uptime proof that reaches the network
happens after 30 minutes (or 60 without the previous commit).  A bad
timing of restarts can then lead to a deregistration as a result (for
example, restarting after 45 minutes, then again after another 45
minutes -- the first uptime proof might not go out for 2:30, which would
cause a deregistration).

It isn't entirely clear why the proof doesn't go out (and it also
doesn't *always* fail to go out--but did in 4 out of 5 tests I did, all
done when an uptime proof should have been accepted by the network--by
being well past the 30 minute minimum).  At a guess, it may be happening
before p2p connections to other nodes are established.  Delaying the
initial proof transmission until 60 seconds of uptime seems to be
considerably more robust in terms of getting it out, and is working for
me consistently when it should (i.e. on restart when the last proof was
>30 minutes).