This should address #353
Because the code would try to resolve a given host as IPv6 and
then fail if it resolved but couldn't connect, nodes which could
resolve IPv6 but did not have IPv6 connectivity to that host for
whatever reason could not connect at all.  This change both makes
IPv4 resolution and connection attempt come first, and also makes
the client attempt IPv6 resolution/connection if either fails for
IPv4.