Fix for v3 proofs not being accepted by v2 nodes
Current 2.0.x nodes will only accept a v10 network uptime proof if the
proof's snode_version_major is exactly equal to 2 (rather than >= 2), so
the proofs generated by a 3.0.x node won't be accepted by 2.0.x nodes.

This commit fakes the snode version put into an uptime proof prior to
the to the v11 fork to a fictitious 2.3.x version instead of the actual
3.0.x to keep v2 nodes happy with the proof.

(The same issue isn't present for future upgrade: the code in 3.0
properly recognizes >= 3 as valid for sending v11 network proofs).