Add last announced SN version to stored proof info
This changed the stored proof info to a struct storing both the
timestamp and the major/minor/patch versions that were included in the
last uptime proof.

The version gets exposed as a "service_node_version" 3-int array in the
get_service_nodes RPC call.

This commit also includes a minor non-critical fix to the uptime proof
pruning as part of the changes: deleting while iterating isn't
guaranteed to iterate through all elements before C++14.