More wallet performance by caching contributed SNs
This caches a filtered-by-contributor list of service nodes in the
node_rpc_proxy so that when the wallet balance requests it the vast
majority of nodes will be only checked (and not included) for the first
input checked, massively speeding up following inputs which only need to
check the service nodes where the account has actually contributed.

This makes the `balance` command on my Raspberry Pi wallet return the
result essentially instantly (I can perceive no delay at all), while
before this it was around half a second (and before the previous
performance commit, around 10s on the same wallet).