Endian fixes and code simplification
This replaces some memcpy's that would have done the wrong thing on a
big-endian system, plus some that were doing the right thing with a nice
simplification.

It adds a tools::memcpy_le() function to do the heavy lifting as well,
which takes a list of simple objects and memcpy's them all into a
std::array<char,N> that it returns to you, doing endian conversion to
little-endian on any integer types.

This greatly simplifies the memcpy code, plus takes care of the endian
issues behind the scenes.