Avoid reinterpret_cast undefined behaviour
Using the resulting pointer of a `reinterpret_cast<void *>` is undefined
behaviour (even though it works on most compilers), but moreover is
unnecessary: all pointers are implicitly convertible to void pointers
and the result of that implicit conversion is *not* undefined behaviour.
(Thus a `static_cast<void *>` is also acceptable, but pointlessly
verbose).

Downcasting via reinterpret_cast is never valid; when type safety is
externally assured, as in this case, static_cast<Derived *> is the
correct operator.