Avoid vla and potential stack overflow
VLAs (which are usually just a wrapper around alloca) are evil in
general, and can easily stack overflow if the number of transactions
gets too large (though this is probably mostly theoretical).  Replace
with a version that uses a local stack 4096-byte buffer that can handle
up to 256 transactions and falls back to dynamic memory allocation for
larger cases (which will be rare).