Remove pointless struct misalignments
The code was littered with `#pragma pack(1)` calls, but these were being
highly misused: they not only pack the data (avoiding padding), but they
*also* change the alignment to byte alignment which means you are likely
to end up with misaligned data access causing performance issues and,
especially on non-x86, memory access failures.

This was a nasty hack needed to get epee to accept a type as a span of
bytes.  But what epee really needed here was *not* that a type has byte
alignment, but rather that it has not member padding.  To this end I
added a specializable struct to epee to allow a type to be specialized
when the caller knows it has no padding (the default implementation
maintains the 1-byte-alignment requirement, which is really the only
generic way to ensure a type contains no padding).

Almost all of them are also *pointless*: the vast majority of them were
around data structures that were *already* fully packed because of the
data sizes involved and so ended up doing nothing useful and only
potentially harming data access.

This also changes the ec_point/ec_scalar alignment (which was 1 anyway,
even without the pack pragma) to be the same as size_t: these types have
a std::hash implementation that reinterprets the hash value as a size_t
for hash bucket selection -- but without size_t alignment that is again
invalid.

The single pack that I kept is in one piece of code where the data
members underlying types consist entirely of 1-byte elements (uint8_ts
or chars) but the packed layout is specifically needed to calculate the
checksum.

Finally there's a fix included here for `blk_checkpoint_header` to use
uint64_t rather than `size_t` so that the size is well-defined
(otherwise it seems the lmdb storage would be different on architectures
with different size_t's, e.g. x86 and amd64).