Data structure & C++ optimizations
Various efficiency improvements to the handling of service node list
data:

- Use an unordered_map for the key image blacklist so that we don't have
to linearly scan every time a key image check is needed.
- getting blacklisted key images was a segfault waiting to happen: it
returned a reference into the transient state to public code, which can
change behind the scenes invalidating the reference.
- print MERROR if a blacklisted key image rollback tries to re-add an
already-blacklisted key image (this should be impossible).
- store transient service node info in a shared_ptr; this reduces the
number of places the data has to be copied, which seems worthwhile as it
is a fairly large struct.
- Use emplace rather than insert to avoid double-copying (or
construct-then-move) of inserted data and elements
- Fix non-atomicity in winner and portions code that could result in
orphaned blocks.  Currently selecting the winner and getting the
portions are two separate calls, so if an update happens between the two
calls, the portions list could be for a different block or winner than
the select winner.  This combines them into a single atomic call.
- various type/enum changes (missing enum use, better static_casts)
- make public keys ordered (lexicographically) so that they can simply
be sorted using the default sort ordering.  (This also lets them be used
in an ordered std::map<public_key, whatever> with needing a custom
compare function, though that isn't currently being used here.)