Add --backup-service-node flag to reduce proof timeout
This is designed to allow people to still run backup service nodes
*without* a race condition between the main and backup SNs that triggers
an undesirable and unnecessary service node IP switch on the network.

A SN running in backup mode won't submit uptime proofs until it has gone
1h45m without seeing an uptime proof on the network for the SN with the
same pubkey in the past 1h45m.  (Which means it also has to have been
alive for at least 1h45m, and the SN must have been registered for ~1h45
worth of blocks).