Tweak decommission credit rules
This starts everyone with 2 hours of credit at deregistration (instead
of none), and reduces the minimum required for decommissioning to 2
hours so that everyone can get some decommission right away.

Otherwise the credit accumulation rate and maximum remain unchanged: 24
blocks per day = 24 hours worth of blocks per 30 days, and a maximum
accumulated credit of 24 hours (which you'll now hit 27.5 days after
registration instead of 30 because of the initial 2 hour credit).