Add a `state_height` to RPC
The field contains the latest height at which the SN status changed:
that is, the height at which the last contribution was received (if
awaiting); the height of activation (if active and never
decommissioned); the height of decommissioning (if currently
decommissioned); or the height of last recommissioning if applicable.

This is need by the block explorer to be able to report useful
information on the current status.