Make cosmetic changes for RPC declarations for the documentation generator
The doc generator doesn't know how to parse using statements, so upgrade
contributor/contributions to global scope in the RPC definitions file so
it's able to parse it for both service node RPC calls.

The best option will be to eventually deprecate the various ways to
query the network down to 1 call that doesn't duplicate functionality.