Check checkpoints for all alt blocks along the alt chain
If you receive a historical checkpoint on your canonical chain that
conflicts with the current blocks, when receiving a new block on the
alternative chain that matches the checkpoints, we should enumerate
along that alt chain how many blocks match checkpoints for reorganizing
to.

Otherwise, upon receiving alt blocks, we only check the latest block
despite having possibly received new checkpoints for blocks in the past.