Remove crypto::crypto_ops
This "class" is a hot mess of uselessness:

- It's defined as a class that has declared constructor/destructor/copy
  assignment, but can't be instantiated (because the declared
  constructors/destructors/etc. aren't defined anywhere).
- There's a bunch of completely useless `friend` declarations in the
  crypto wrappers (public_key, etc.) that do absolutely nothing since
  those wrappers don't *have* anything private in them.
- Every (private) static function of crypto_ops is completely duplicated
  outside crypto_ops taking exactly the same arguments.
- In order to call all those private static functions the containing
  namespace has inlined free functions calling the private static
  functions inside crypto_ops with identical arguments.
- Because the above isn't allowed, all of the definitions are repeated
  a *third* time with friend function declarations inside crypto_ops.

So basically everything inside crypto_ops is exactly callable outside
crypto_ops with the same name and same arguments, but has this
completely pointless layer of indirection that adds nothing for no
reason.

This commits rips out all this useless crap and just makes the various
previous-crypto_ops static functions plain functions in the `crypto`
namespace (which is the only place they can be used and called anyway).