Generate ed25519 and x25519 keys for a service node
The Monero pubkey/privkey is pretty much useless outside of Monero: we
can't use it with standard tools.  This adds two new keypairs: a ed25519
(usable for signing) and a x25519 (usable for encryption).

The x25519 are converted from the ed25519 (and so only the ed25519
pubkey needs to be sent out to the network).

It would theoretically be possible to derive all from the monero pubkey,
but I was advised not to (in particular for avoiding issues in any
potential future audit).

Quorumnet, in particular, needs the x25519 for authenticating/encrypting
communication, but other devs have expressed an interest in also having
a standard ed25519 available, hence the double keys.