Remove hardcoded checkpoint on testnet
Hardcoded checkpoints are not expected after HF13. At some point during
the previous releases's testing period we added a hardcoded checkpoint
because we redid the hardforking blocks on testnet due to some
incompatible changes- but were tenable to reboot the network on a small
network like the testnet.

This commit deletes the checkpoint instead of adding an exception so as
to not require letting the service node list accept hardcoded
checkpoints after HF13, or worse, if somehow a corrupt checkpoint was
added to the DB somehow, and on retrieval we return a checkpoint with no
signatures, this could sneak through as it'd be detected as a hardcoded
checkpoint once loaded from the DB.

We could provide an exception if it was a checkpoint on height 127XXX
and it was testnet, but this is too intrusive change to the conditional
for an insignifcant testing change.