Eliminate superfluous new top block info logs
This eliminates superfluous info logs:

    Sync data returned a new top block candidate: blah blah

Blink changed the logic here to not early return when we already have
the remote's top block because we may need to sync blinks even if the
top block is the same, but then resulted in lots of the above info
messages whenever we sync up with a new peer (even if the peer has no
new blocks for us).

This restores it to the 5.x behaviour where we only want to display the
info if we don't have the remote's block.