Remove NO_AES option
It is dumb: the invoker can specify whether they want -maes added based
on the -march value they provide or CXXFLAGS that they set; we should
not be forcing it on through an obscure option.

It looks like the only reason this option was added was because the CN
turtle devs didn't know that there is a __AES__ define that gets defined
when you are compiling with AES optimizations.  Removing this option
makes the build work properly on old non-AES-supporting CPUs where the
default ARCH=native will now do the right thing and disable the AES
instruction codepath.