Flatten epee handle_http_request callstack
Remove CHAIN_HTTP_TO_MAP2 macro as it's used for templatizing on
command_type. But also, the preamble in CHAIN_HTTP_TO_MAP2 repeats the
the same setup as it does in http_protocol_handler.h before it is
dispatched into Monero land.

So instead move the extra setup steps (logging and error handling) into
epee code. handle_http_request is now implemented in Monero land and
only concerns, actually handling the http request.

This also allows us to put the RPC handling code in the header to
the cpp file and improve debugability.