Pulse: Fix incorrect validator quorum selection
We don't want to choose from [0, (partition_index - running_it)) from
the active Service Node list. What we desire is a sliding window that
always considers half of the list.

Once we swap a Service Node into place, 1 service node from the second
half of the list will enter the first half. Otherwise, if we have
exactly PULSE_QUORUM_SIZE nodes to choose from, we would never be able
to fulfill a quorum as there are only half of the nodes to choose from,
(the other half would be in the 2nd half of the list and never
be considered) in the old scheme.