Remove -Ofast, use cmake defaults for release/debug
-Ofast turns on -ffast-math, which allows gcc to do IEEE-violating FP
math optimizations.  The consequence of this is that it means a Release
build and a Debug build produce *different* difficulty values.  (It may
also have contributed to difficulty value divergence).

Turn it off because we are not Gentoo.

There is actually little point of setting CMAKE_CXX_FLAGS_RELEASE at all
here: adding `-DNDEBUG` is already the default for a cmake release
build, and cmake similarly has a release build default optimization
level (which I think is `-O3` -- which is already what we're doing with
`-Ofast` on Linux, except that `-Ofast` turns on unsafe stuff on top of
`-O3`).

For Debug build, this removes -Og and lets cmake use its default (-O0),
which is usually better when developing: lack of optimization means
faster binaries, are easier to trace.  The downside is enormous binaries
(~250MB), but that seems at least manageable for a debug build.