Use integer math for difficulty calculation
This replaces the floating point math difficulty calculation with
integer math.

It is still a work in progress: this calculation appears to be produce
individual block difficulty values that are *slightly* too small by
somewhere around 0.05% - 0.1% (over the entire chain on mainnet,
currently, I get a *cumulative* difficulty of:

Height: 560788 prev difficulty: 4218954063635982, new difficulty: 4215801479958170

which is 99.925% of the proper value, suggesting that this small
deviation in the integer calculation is fairly consistent over the
entire chain (and is likely caused by the fact that the intermediate
integer division calculations truncate).