Recommission credit
This changes the credit rules for recommission: instead of being
restored with no credit at all, with this you get restored with 2 blocks
of credit lost (up to all of your credit) for each block of decommission
time you had.

For example, if you had 1000 blocks of credit and are decommissioned for
150 blocks, you'll get recommissioned with 700 blocks of credit.  If you
were decommissioned for 500 or more blocks then you'll get
decommissioned with 0 blocks of credit.

This applies the effects immediately rather than waiting for the hard
fork.  Because this isn't a consensus-breaking change it seems easier to
not worry about the HF rules.  The consequences are that the rules that
get applied to determining how much credit you have will depend on the
nodes involved in testing up until the hard fork: you *may* be subject
to the harsher v7 rules, or you may fall under v8 rules if enough of the
testing quorum has upgraded, but you'll never end up worse than the v7
rules.