Fix internal RPC requests for service nodes
Fixes a couple of issues that caused problems when making requests to
GET_SERVICE_NODES:

- Internal requests making an rpc request for `GET_SERVICE_NODES::req{}`
  weren't ending up specifying `"all":true` in requested_fields *or* any
  other fields, so they requested no fields and got no fields in
  response.

That, however, exposed another problem:

- the serialized value *loading* of service node entries shouldn't be
  looking at requested fields at all: it should just deserialize
  whatever is incoming.  This was resulting in a parsed incoming service
  node RPC request ending up with all-empty data even when the data
  *was* there, which consequently broke the get-all-service-nodes call
  in the wallet, making the wallet think it had no locked service nodes.

The fix for issue 1 is to ignore requested_fields when `!is_store`.

The fix for issue 2 is to make requested_fields a std::optional and
handle the null-mean-all logic in the RPC call itself rather than in the
serialization layer.