Fix max uint64_t errors when contributing to a reserved spot
When trying to fill a reserved spot of a service node that is fully
reserved the wallet is getting back the UINT64_MAX error code value for
min_contrib_total, but not properly handling it resulting in:

Error: You must contribute at least 18446744073.709551615 loki to become a contributor for this service node.

Fix it to recognize the error value when filling a reserved spot.