Fix JSON serialization of pulse block data
There was a missing call to nested the values in an object, so the
resulting json ended up as:

    "pulse":,"random_value":"b01b8d1290cbac40917a23a08b343bec","round":0,"validator_bitset":2047,

instead of:

    "pulse":{"random_value":"b01b8d1290cbac40917a23a08b343bec","round":0,"validator_bitset":2047},

Modernize it to avoid macros, and fix it by beginning an object.

(This won't affect binary serialization; object serialization only has
an effect on json serialization and is a no-op for binary
serialization).