get_block_template: Allow unaccounted dust in construct_miner_tx
- On some of the SN's who are queued for payout, requesting a block
template with it fails because we end up paying less than the total
amount allocated.

This was previously solved by having a `service_node_paid` in
`reward_parts` that I removed to simplify `reward_parts`.

This is only a client side sanity check and does not affect consensus.
The rest of the codebase uses the original calculate_sums_of_portions on
the total allocated reward (which when it does the division out, the
remainder is not included in the payouts and the verifying code expects
that, so everything should continue to work fine).